requirements.txt
Flask==2.2.2
gunicorn==20.1.0
python-dotenv==0.21.1
pytest==7.0.1
https://download.pytorch.org/whl/cpu/torch-2.0.0%2Bcpu-cp39-cp39-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.15.1%2Bcpu-cp39-cp39-linux_x86_64.whl
segmentation-models-pytorch==0.3.2
requests-toolbelt==0.10.1
Tensorflow-cpu==2.12.0